Revert "Vector db init"
