--- requirements.txt
@@ -1,143 +1,0 @@
-aiohappyeyeballs==2.6.1
-aiohttp==3.11.14
-aiosignal==1.3.2
-annotated-types==0.7.0
-annoy==1.17.3
-anyio==4.9.0
-appdirs==1.4.4
-asttokens==3.0.0
-attrs==25.3.0
-blinker==1.9.0
-boto3==1.37.21
-botocore==1.37.21
-cachetools==5.5.2
-certifi==2025.1.31
-cffi==1.17.1
-charset-normalizer==3.4.1
-click==8.1.8
-colorama==0.4.6
-contourpy==1.3.1
-cryptography==44.0.2
-cycler==0.12.1
-dataclasses-json==0.6.7
-datasets==3.5.0
-decorator==5.2.1
-dill==0.3.8
-dnspython==2.7.0
-executing==2.2.0
-fashion-clip==0.2.2
-filelock==3.18.0
-filetype==1.2.0
-Flask==3.1.0
-flask-cors==5.0.1
-fonttools==4.56.0
-frozenlist==1.5.0
-fsspec==2024.12.0
-google-ai-generativelanguage
-google-api-core==2.24.2
-google-api-python-client==2.166.0
-google-auth==2.38.0
-google-auth-httplib2==0.2.0
-google-generativeai
-googleapis-common-protos==1.69.2
-greenlet==3.1.1
-grpcio==1.71.0
-grpcio-status==1.71.0
-grpcio-tools==1.71.0
-h11==0.14.0
-h2==4.2.0
-hpack==4.1.0
-httpcore==1.0.7
-httplib2==0.22.0
-httpx==0.28.1
-httpx-sse==0.4.0
-huggingface-hub==0.29.3
-hyperframe==6.1.0
-idna==3.10
-ipyplot==1.1.2
-ipython==9.0.2
-ipython_pygments_lexers==1.1.1
-itsdangerous==2.2.0
-jedi==0.19.2
-Jinja2==3.1.6
-jmespath==1.0.1
-jsonpatch==1.33
-jsonpointer==3.0.0
-kiwisolver==1.4.8
-langchain==0.3.22
-langchain-community==0.3.20
-langchain-core==0.3.49
-langchain-google-genai
-langchain-text-splitters==0.3.7
-langsmith==0.3.21
-lib==4.0.0
-MarkupSafe==3.0.2
-marshmallow==3.26.1
-matplotlib==3.10.1
-matplotlib-inline==0.1.7
-mpmath==1.3.0
-multidict==6.2.0
-multiprocess==0.70.16
-mypy-extensions==1.0.0
-networkx==3.4.2
-numpy==2.2.4
-orjson==3.10.16
-packaging==24.2
-pandas==2.2.3
-parso==0.8.4
-pillow==11.1.0
-portalocker==2.10.1
-prompt_toolkit==3.0.50
-propcache==0.3.1
-proto-plus==1.26.1
-protobuf==5.29.4
-pure_eval==0.2.3
-pyarrow==19.0.1
-pyasn1==0.6.1
-pyasn1_modules==0.4.2
-pycparser==2.22
-pydantic==2.11.0
-pydantic-settings==2.8.1
-pydantic_core==2.33.0
-Pygments==2.19.1
-PyJWT==2.10.1
-pymongo==4.11.3
-pyparsing==3.2.3
-python-box==7.3.2
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
-pytz==2025.2
-pywin32==310
-PyYAML==6.0.2
-qdrant-client==1.13.3
-regex==2024.11.6
-requests==2.32.3
-requests-toolbelt==1.0.0
-rsa==4.9
-s3transfer==0.11.4
-safetensors==0.5.3
-setuptools==78.1.0
-shortuuid==1.0.13
-six==1.17.0
-sniffio==1.3.1
-SQLAlchemy==2.0.40
-stack-data==0.6.3
-sympy==1.13.1
-tenacity==9.0.0
-tokenizers==0.21.1
-torch==2.6.0
-tqdm==4.67.1
-traitlets==5.14.3
-transformers==4.50.2
-typing-inspect==0.9.0
-typing-inspection==0.4.0
-typing_extensions==4.13.0
-tzdata==2025.2
-uritemplate==4.1.1
-urllib3==2.3.0
-validators==0.34.0
-wcwidth==0.2.13
-Werkzeug==3.1.3
-xxhash==3.5.0
-yarl==1.18.3
-zstandard==0.23.0
