--- src/AI/model_service/api/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-api_blueprint = Blueprint('ai', __name__)
-
-from . import handle_prompt
