--- src/AI/model_service/fashion_search/__init__.py
@@ -1,3 +1,0 @@
-from .image_to_image import ImageToImageSearch
-from .image_to_text import ImageToTextGenerator
-from .text_to_image import TextToImageSearch
