--- src/AI/model_service/utils/__init__.py
@@ -1,1 +1,0 @@
-from .decode_base64_image import decode_base64_image
