--- src/AI/model_structure/utils/__init__.py
@@ -1,1 +1,0 @@
-from .extract_images import extract_all_images
